client4: do just enough to get it to build
